requirements.txt
Django>=2.1.3,<3
psycopg2-binary==2.7.6
PyJWT>=1.6.4,<2
